refactor: move authentication methods and refactor imports across the application
