+++ core/src/app/(protected)/app/(sidebar)/notifications/page.tsx
@@ -1,10 +1,7 @@
-import { headers } from "next/headers";
-
-import { getAuthDataOrRedirectToLogin } from "@/zap/lib/auth/redirects";
+import { getAuthServerDataOrRedirectToLoginService } from "@/zap/auth/services";
 
 export default async function NotificationsPage() {
-  const _headers = await headers();
-  await getAuthDataOrRedirectToLogin(_headers);
+  await getAuthServerDataOrRedirectToLoginService();
 
   return <></>;
 }
