+++ core/src/app/(protected)/app/billing/success/page.tsx
@@ -1,8 +1,7 @@
-import { headers } from "next/headers";
 import Link from "next/link";
 
+import { getAuthServerDataOrRedirectToLoginService } from "@/zap/auth/services";
 import { ZapButton } from "@/zap/components";
-import { getAuthDataOrRedirectToLogin } from "@/zap/lib/auth/redirects";
 
 interface BillingSuccessPageProps {
   searchParams: Promise<{
@@ -13,8 +12,7 @@ interface BillingSuccessPageProps {
 export default async function BillingSuccessPage({
   searchParams,
 }: BillingSuccessPageProps) {
-  const _headers = await headers();
-  await getAuthDataOrRedirectToLogin(_headers);
+  await getAuthServerDataOrRedirectToLoginService();
 
   const { checkout_id } = await searchParams;
