+++ core/src/app/(public)/(legal)/cookie-policy/page.tsx
@@ -1,5 +1,5 @@
-import { LegalPage } from "@/zap/components/legal/legal-template";
-import { generateLegalMetadata } from "@/zap/lib/legal/utils";
+import { LegalPage } from "@/zap/legal/pages";
+import { generateLegalMetadata } from "@/zap/legal/utils";
 
 const SLUG = "cookie-policy";
