+++ core/src/app/(public)/blog/layout.tsx
@@ -1,5 +1,5 @@
-import { Footer } from "@/zap/components/common/footer";
-import { Header } from "@/zap/components/common/header";
+import { Footer } from "@/zap-old/components/common/footer";
+import { Header } from "@/zap-old/components/common/header";
 
 interface BlogLayoutProps {
   children: React.ReactNode;
