+++ core/src/hooks/rpc/use-example.ts
@@ -1,8 +1,8 @@
 "use client";
 import "client-only";
 
-import { useZapQuery } from "@/zap/lib/api/hooks/use-zap-query";
-import { orpc } from "@/zap/lib/orpc/client";
+import { useZapQuery } from "@/zap/api/hooks";
+import { orpc } from "@/zap/api/providers/orpc/client";
 
 export function useExample() {
   return useZapQuery(orpc.example.queryOptions());
