+++ core/src/rpc/middlewares.ts
@@ -1,2 +1,2 @@
-export * from "@/rpc/middlewares/auth.middleware";
 export * from "@/rpc/middlewares/base.middleware";
+export * from "@/zap/auth/rpc/middlewares";
