+++ core/src/zap/blog/data/nextjs.mdx
